Add files via upload
In this code, the fuzzer will iterate through the defined endpoints and send requests to each of them. If a 404 response is received for any of the endpoints, it will print the notification and continue fuzzing the remaining endpoints. If the response is not a 404, it will print the JSON data and endpoint information.